Move location-related code into location.js
